Aubatch.c split into multiple source files.
